requirements-base.txt
fastapi~=0.104.1
pydantic~=2.5.1
python-jose~=3.3.0
mongoengine~=0.27.0
Pillow~=10.1.0
passlib~=1.7.4
python-dotenv~=1.0.0
phonenumbers~=8.13.25
telesign~=2.2.2
bcrypt~=4.0.1
apscheduler~=3.10.4
pymongo~=4.6.0
python-multipart~=0.0.6
pytz~=2023.3
vincenty~=0.1.4
sse-starlette~=1.6.5
boto3~=1.28.85

